One reachability walk, and dry runs that cost nothing
The prune and bin dry runs were still full-collection reads — ~350 document
reads each to tell you what they WOULD do. That is the clearest case of a read
that need not be live, and together with the audit it is what exhausted the
project's daily quota and took the live app down.

Both now take --offline and answer from the snapshot the last live run wrote.

--offline WITH --apply IS REFUSED, in both. That guard is the point of the
change rather than a detail: --offline is valuable precisely because the data
is old, and --apply decides what to destroy from exactly the data it just read.
A delete computed from a stale snapshot would act on a graph that has since
moved — removing something re-filed in the meantime, or missing something newly
orphaned. Reading stale saves quota; writing stale is a bug with the same shape
as the accident these scripts exist to clean up after. It exits 2 and names the
live command.

The shared parts moved to scripts/timeline-snapshot.mjs. All three scripts had
their OWN copy of the reachability walk, clipsOf, childIdsOf and credential —
three chances for them to disagree about what "reachable" means, which is
exactly the disagreement that would have one script delete what another
considers live. The dangling-reference handling added to the audit yesterday
existed in only one of the three copies, so the other two were still counting
documents that do not exist as reachable.

Verified against a fixture snapshot: the audit's output is byte-identical to
before the refactor; prune sees 1 deletable and 1 keeping; bin plans 2; and
both --offline --apply attempts refuse with exit 2. The live read/write paths
are unchanged and still need a real scan to exercise.
